+++ packages/select/src/interface.ts
@@ -34,10 +34,11 @@ export type SelectValue =
 export interface SelectProps<T extends SelectValue = SelectValue>
   extends Omit<
       HTMLAttributes<HTMLDivElement>,
-      "defaultValue" | "onChange" | "prefix" | "onFocus" | "onBlur"
+      "defaultValue" | "onChange" | "prefix" | "onFocus" | "onBlur" | "onSelect"
     >,
     BoxProps {
   autoAlignPopupWidth?: boolean
+  inputAsOption?: boolean
   colorScheme?: SelectColorScheme
   allowClear?: boolean
   variant?: SelectVariant
@@ -61,6 +62,9 @@ export interface SelectProps<T extends SelectValue = SelectValue>
   filterOption?:
     | boolean
     | ((inputValue: string | number, option: SelectOptionObject) => boolean)
+  defaultFilterOption?:
+    | boolean
+    | ((inputValue: string | number, option: SelectOptionObject) => boolean)
   onChange?: (value?: T) => void
   onClear?: () => void
   readOnly?: boolean
@@ -70,6 +74,7 @@ export interface SelectProps<T extends SelectValue = SelectValue>
   dropdownProps?: DropdownProps
   options?: SelectOptionObject[] | string[] | number[]
   onDeselect?: (value: string | number | SelectOptionObject) => void
+  onSelect?: (value: string | number | SelectOptionObject) => void
   multiple?: boolean
   onBlur?: (e: SyntheticEvent) => void
   onFocus?: (e: SyntheticEvent) => void