+++ packages/breadcrumb/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @illa-design/breadcrumb
 
+## 1.1.0
+
+### Minor Changes
+
+- update design system
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @illa-design/dropdown@1.1.0
+  - @illa-design/icon@1.1.0
+  - @illa-design/system@1.1.0
+  - @illa-design/theme@1.1.0
+
 ## 1.0.31
 
 ### Patch Changes