+++ packages/input-tag/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @illa-design/input-tag
 
+## 1.1.0
+
+### Minor Changes
+
+- update design system
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @illa-design/icon@1.1.0
+  - @illa-design/input@1.1.0
+  - @illa-design/space@1.1.0
+  - @illa-design/system@1.1.0
+  - @illa-design/tag@1.1.0
+  - @illa-design/theme@1.1.0
+
 ## 1.0.31
 
 ### Patch Changes