+++ packages/popconfirm/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @illa-design/popconfirm
 
+## 1.1.0
+
+### Minor Changes
+
+- update design system
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @illa-design/button@1.1.0
+  - @illa-design/config-provider@1.1.0
+  - @illa-design/icon@1.1.0
+  - @illa-design/space@1.1.0
+  - @illa-design/system@1.1.0
+  - @illa-design/theme@1.1.0
+  - @illa-design/trigger@1.1.0
+  - @illa-design/typography@1.1.0
+
 ## 1.0.31
 
 ### Patch Changes